+++ neoforge/src/main/java/com/buuz135/modderspiggybank/NeoForgeModdersPiggyBankMod.java
@@ -9,6 +9,7 @@ import net.neoforged.neoforgespi.language.IModInfo;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 
 @Mod(Constants.MOD_ID)
 public class NeoForgeModdersPiggyBankMod {
@@ -16,7 +17,13 @@ public class NeoForgeModdersPiggyBankMod {
     public NeoForgeModdersPiggyBankMod(IEventBus eventBus) {
         var authorList = new HashMap<String, AuthorInformation>();
         for (IModInfo mod : ModList.get().getMods()) {
-            var authors = ((String) mod.getConfig().getConfigElement("authors").orElse("")).split(",");
+            var dirty = mod.getConfig().getConfigElement("authors").orElse("");
+            var authors = new ArrayList<String>();
+            if (dirty instanceof String dirtyString) {
+                authors.addAll(List.of(dirtyString.split(",")));
+            } else {
+                authors = (ArrayList<String>) dirty;
+            }
             for (String author : authors) {
                 if (author.isEmpty()) continue;
                 authorList.computeIfAbsent(author.trim(), string -> new AuthorInformation(string, new ArrayList<>())).modNames().add(mod.getDisplayName());