+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+# 1.0.1
+* Fixed special case where mods would use an array in their author data
+* Added custom main menu screen
+* Mod list only scrolls if there are more than 3 mods
+* Added support for PackMenu
+
 # 1.0.0
 * First release